Python/DjangoTests/loja_de_carros/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-static==0.0.6
Django==3.0.5
django-bootstrap4==1.1.1
django-stdimage==5.0.3
gunicorn==20.0.4
Pillow==7.1.1
progressbar2==3.50.1
python-utils==2.4.0
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
